requirements.txt
altair==5.2.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
imagecodecs==2023.9.18
imageio==2.33.1
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy
opencv-python-headless 
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
streamlit-image-comparison==0.0.4
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
tifffile==2023.12.9
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.227
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
